Minor fixes to update_tv_show_file_names_thread.py
Added some minor fixes to update_tv_show_file_names_thread.py. The string name for the tv show was not formatted correctly. Also added in a TODO task in the model.py to have the program detect if the media files have already been updated. If so, inform the user given Version-1 of this app does not update media files that have already been updated.