bug fix -- better media file check
Creating a fuction to do a more indepth job of checking if the media file(s) in the selected directory have been updated or not. For the way it is now, the program can be fooled if only a handful of the files have been updated for a TV show. If it's the case, need to inform the user of the situation.