Connected signals from create_media_folder_thread.py
Finished connecting the signals from create_media_folder_thread.py and back. Learned that signals can be "chained" together, thus don't need a function to link to signals together.
Updated some names of the functions, classes, and variables to be from "Create Media Folders" to "Create Media Folder". Reason why is the function will be for creating a brand new Media Folder and not also for adding new folders to an existing Media Folder. That will be a new button.