Restructured correct_media_file_format.py
Completely changed correct_media_file_format.py. Changed it from a class to a function and simplified it to where the files in the extra folder no longer need a specific file format.
Need to pass another variable into the scan_directory_thread.py to enable/disable the correct_media_file_format.py to return files in extra folders. This will require updating the GUI to present a popup to ask the user if they wish to grab the extra folder files or not.