passed more values into scan_directory_thread.py
Added two more values to be passed into scan_directory_thread.py--current_media_file_list and files_in_extra_folders_are_formated. also added logic to check to see if the media file was already found in a previous scan.
The next step is to present these values to the user and allow user to make changes if necessary.