auto_update_media_files_thread.py work
Finished creating the logic for the auto_update_media_files_thread.py. Next step is to create a function to update the media file(s) that aren't formated correctly.

For the model.py, created functions to act as "slots" for threads to couple to before being sent onwards. Useful for debugging purposes.

Changed up the correct_media_file_format.py to have a class to hold the different patterns to check against files/folders, and created a new function who's sole purpose is to check the files in a single media folder. The class will be useful when it comes time for the manual media file update feature.