working on automatic_media_file_update.py
Finished creating the logic to prep the files for proper file format for movies and extra folders. Working on the TV show episodes.

Added more methods to the MediaFile class to help with the process of getting the files to the proper format.

Continuing to modify/add hance the FolderAndFilePatterns class. Now it is possible to access the regex expressions if necessary.