autocorrect file format complete
Finished creating the code that will automatically generate the correct file format for the media file based on the folder it is in.

Modified the FolderAndFilePatterns class to include groups within the tv_episode_file_format regex pattern.

Fixed incorrect file naming for the movie folder auto file format, and did a f"" instead of concatenating the string together. Makes the code look cleaner and makes it easier to identify the different parts.