Linked messagebox to model for task completed
Made a signal from the auto_update_media_files_thread.py to the model.py that activates another signal to the main_window.py to tell view.py to activate a messagebox to inform the user the task is complete.

Also changed func names of the pre-existing slots the threads connect to on the model.py side.