changed logic and flow of manual update
Made some changes how the manual update process works. Instead, have the program send the data to check to see if any of the rename file names match each other and if that passes then send it off to be updated. Reason why is if the duplication test fails, the window stays open allowing the user to make the necessary changes.