dynamic button enabling for media folder creation window
Made it to where when the user clicks the button to create a new media folder it will disable all buttons while the process is being worked on. Once process is down the buttons are re-enabled and the window stays open so the user can create another media folder if they desire.