dynamic button enabling for modified media folder
Made it to where when the user clicks the button to modified existing media folder it will disable all buttons while the process is being worked on. Once process is down the buttons are re-enabled and the window stays open to allow user to modify another media folder if they wish.